// packages/ocpi-base/src/model/Facilities.ts
// SPDX-FileCopyrightText: 2025 Contributors to the CitrineOS Project
//
// SPDX-License-Identifier: Apache-2.0

export enum Facilities {
  HOTEL = 'HOTEL',
  RESTAURANT = 'RESTAURANT',
  CAFE = 'CAFE',
  MALL = 'MALL',
  SUPERMARKET = 'SUPERMARKET',
  SPORT = 'SPORT',
  RECREATION_AREA = 'RECREATION_AREA',
  NATURE = 'NATURE',
  MUSEUM = 'MUSEUM',
  BIKE_SHARING = 'BIKE_SHARING',
  BUS_STOP = 'BUS_STOP',
  TAXI_STAND = 'TAXI_STAND',
  TRAM_STOP = 'TRAM_STOP',
  METRO_STATION = 'METRO_STATION',
  TRAIN_STATION = 'TRAIN_STATION',
  AIRPORT = 'AIRPORT',
  PARKING_LOT = 'PARKING_LOT',
  CARPOOL_PARKING = 'CARPOOL_PARKING',
  FUEL_STATION = 'FUEL_STATION',
  WIFI = 'WIFI',
}